Refactor file/class location
