+++ ActiveDirectoryQuerier.Tests/ComboBoxParameterViewModelTests.cs
@@ -2,6 +2,7 @@ using System.Collections.ObjectModel;
 using System.Management.Automation.Runspaces;
 using ActiveDirectoryQuerier.ActiveDirectory;
 using ActiveDirectoryQuerier.PowerShell;
+using ActiveDirectoryQuerier.ViewModels;
 
 namespace ActiveDirectoryQuerier.Tests;
