+++ ActiveDirectoryQuerier/ActiveDirectoryQuerier.csproj
@@ -25,8 +25,4 @@
 		<PackageReference Include="Microsoft.PowerShell.SDK" Version="7.2.18" />
 	</ItemGroup>
 
-	<ItemGroup>
-	  <Folder Include="ViewModels\" />
-	</ItemGroup>
-
 </Project>
