+++ ActiveDirectoryQuerier/MainWindowViewModel.cs
@@ -10,11 +10,11 @@ using System.Windows.Input;
 using ActiveDirectoryQuerier.ActiveDirectory;
 using ActiveDirectoryQuerier.PowerShell;
 using ActiveDirectoryQuerier.Queries;
+using ActiveDirectoryQuerier.ViewModels;
 using Microsoft.Win32;
 
 namespace ActiveDirectoryQuerier;
 
-/// TODO: Place class into ViewModel folder.
 public sealed class MainWindowViewModel : INotifyPropertyChanged
 {
     // [ Fields ] ------------------------------------------------------------------- //
